add noOfQueryMixes

+++ benchmark-service/src/main/java/org/dicegroup/basilisk/benchmarkService/model/iguana/task/Worker.java
@@ -15,6 +15,6 @@ public class Worker {
 
     private Integer fixedLatency;
 
-    private Integer timeOut;
+    private Long timeOut;
 
 }
